+++ test/test-variable-analyzer-parse.cpp
@@ -47,7 +47,7 @@ TEST(Test_Variable_Analyzer_Parse, generic_parameter_use_before_declaration) {
   EXPECT_THAT(v.errors,
               ElementsAreArray({
                   DIAG_TYPE_2_OFFSETS(
-                      &input, Diag_Variable_Used_Before_Declaration,       //
+                      &input, Diag_Cyclic_TypeScript_Type_Definition,      //
                       use, u8"function f<T extends "_sv.size(), u8"T"_sv,  //
                       declaration, u8"function f<"_sv.size(), u8"T"_sv),
               }));